+++ src/middleware/requestLogger.ts
@@ -1,13 +1,21 @@
 import { Request, Response, NextFunction } from 'express';
 import { getTimeStamp, getUserId, handleLog } from '../utils';
-import { TAuditOptions, TFileConfig } from '../types';
+import { TAuditOptions, TFileConfig, TSaveContext } from '../types';
+import { userProfile } from '../utils/user';
 
-export const requestLogger = (config: TAuditOptions, fileConfig: TFileConfig, saveContent: (config: TFileConfig, content: any) => void) => {
+
+export const requestLogger = (config: TAuditOptions, fileConfig: TFileConfig) => {
     return (req: Request, res: Response, next: NextFunction) => {
         const start = Date.now();
+        const route = req.originalUrl;
+        const ip = req.ip ?? "unknown";
+        const userAgent = req.headers['user-agent'] || '';
+        const user = getUserId(req);
+        const id = (typeof user === "string") ? user : user ? user?.id : "unknown";
+
+        userProfile.BuildProfile(id, route, ip, userAgent);
 
         res.on('finish', () => {
-            const user = getUserId(req);
             const duration = Date.now() - start;
 
             if ((res as any)._suppressAudit) {
@@ -18,32 +26,30 @@ export const requestLogger = (config: TAuditOptions, fileConfig: TFileConfig, sa
                     duration,
                     method: req.method,
                     statusCode: req.statusCode || 500,
-                    ip: req.ip ?? "unknown",
-                    route: req.originalUrl,
-                    userAgent: req.headers['user-agent'],
+                    ip,
+                    route,
+                    userAgent,
                     ...(config.useTimeStamp ? { timeStamp: getTimeStamp() } : {}),
                 };
-                handleLog(config, fileConfig, saveContent, content);
+                handleLog(config, fileConfig, content);
                 return;
             }
-
             const content = {
                 type: "request",
                 action: "incoming request",
                 duration,
                 method: req.method,
                 statusCode: req.statusCode || 200,
-                route: req.originalUrl,
+                route: route,
                 statusMessage: req.statusMessage || "success",
-                ip: req.ip ?? "unknown",
-                userAgent: req.headers['user-agent'],
-                userId: (typeof user === "string") ? user : user ? user?.id : "unknown",
+                ip,
+                userAgent,
+                userId: id,
                 ...(config.useTimeStamp ? { timeStamp: getTimeStamp() } : {})
 
             };
-            handleLog(config, fileConfig, saveContent, content);
+            handleLog(config, fileConfig, content);
         });
-
         next();
     };
 };